+++ lib/domain/repositories/compose_repository.dart
@@ -4,10 +4,10 @@ import 'package:horizon/domain/entities/raw_transaction.dart';
 
 abstract class ComposeRepository {
   Future<RawTransaction> composeSend(
-      String sourceAddress, String destination, String asset, double quantity,
+      String sourceAddress, String destination, String asset, int quantity,
       [bool? allowUnconfirmedTx, int? fee]);
   Future<ComposeSend> composeSendVerbose(
-      String sourceAddress, String destination, String asset, double quantity,
+      String sourceAddress, String destination, String asset, int quantity,
       [bool? allowUnconfirmedTx, int? fee]);
   Future<ComposeIssuance> composeIssuance(
       String sourceAddress, String name, double quantity,
